Move web/server under main.ts
