+++ src/server.ts
@@ -19,10 +19,10 @@ import { DB } from './db'
 import { EventsProcessor } from './ingestion/process-event'
 import { startQueue } from './main/queue'
 import { startSchedule } from './main/services/schedule'
+import { startFastifyInstance, stopFastifyInstance } from './main/web/server'
 import { status } from './status'
 import { PluginsServer, PluginsServerConfig, Queue, ScheduleControl } from './types'
 import { createPostgresPool, createRedis, delay, UUIDT } from './utils'
-import { startFastifyInstance, stopFastifyInstance } from './web/server'
 
 const { version } = require('../package.json')
